Refactor ViT model to include attn_implementation and torch_dtype parameters

+++ sc4000/train/trainer/weighted_trainer.py
@@ -13,6 +13,9 @@ class WeightedTrainer(Trainer):
         labels = inputs.get("labels")
         outputs = model(**inputs)
         logits = outputs.get("logits")
-        loss_fct = torch.nn.CrossEntropyLoss(weight=self.weight.to(logits.device), label_smoothing=self.label_smoothing)
+        loss_fct = torch.nn.CrossEntropyLoss(
+            weight=self.weight.to(logits.dtype).to(logits.device),
+            label_smoothing=self.label_smoothing,
+        )
         loss = loss_fct(logits.view(-1, self.model.config.num_labels), labels.view(-1))
         return (loss, outputs) if return_outputs else loss
